+++ lib/flutter_chat_ui.dart
@@ -3,6 +3,7 @@ library flutter_chat_ui;
 export 'src/chat_l10n.dart';
 export 'src/chat_theme.dart';
 export 'src/models/emoji_enlargement_behavior.dart';
+export 'src/models/preview_tap_options.dart';
 export 'src/models/send_button_visibility_mode.dart';
 export 'src/util.dart' show formatBytes, isConsistsOfEmojis;
 export 'src/widgets/attachment_button.dart';