+++ lib/src/models/date_header.dart
@@ -6,12 +6,16 @@ import 'package:meta/meta.dart';
 class DateHeader extends Equatable {
   /// Creates a date header.
   const DateHeader({
+    required this.dateTime,
     required this.text,
   });
 
   /// Equatable props
   @override
   List<Object> get props => [text];
+
+  /// Message date
+  final DateTime dateTime;
 
   /// Text to show in a header
   final String text;