+++ lib/src/widgets/message.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
+import 'package:visibility_detector/visibility_detector.dart';
+
 import '../models/emoji_enlargement_behavior.dart';
+import '../models/preview_tap_options.dart';
 import '../util.dart';
 import 'file_message.dart';
 import 'image_message.dart';
@@ -21,14 +24,18 @@ class Message extends StatelessWidget {
     this.fileMessageBuilder,
     required this.hideBackgroundOnEmojiMessages,
     this.imageMessageBuilder,
+    required this.isTextMessageTextSelectable,
     required this.message,
     required this.messageWidth,
     this.onAvatarTap,
+    this.onMessageDoubleTap,
     this.onMessageLongPress,
     this.onMessageStatusLongPress,
     this.onMessageStatusTap,
     this.onMessageTap,
+    this.onMessageVisibilityChanged,
     this.onPreviewDataFetched,
+    required this.previewTapOptions,
     required this.roundBorder,
     required this.showAvatar,
     required this.showName,
@@ -72,6 +79,9 @@ class Message extends StatelessWidget {
   final Widget Function(types.ImageMessage, {required int messageWidth})?
       imageMessageBuilder;
 
+  /// See [TextMessage.isTextMessageTextSelectable]
+  final bool isTextMessageTextSelectable;
+
   /// Any message type
   final types.Message message;
 
@@ -81,21 +91,31 @@ class Message extends StatelessWidget {
   // Called when uses taps on an avatar
   final void Function(types.User)? onAvatarTap;
 
+  /// Called when user double taps on any message
+  final void Function(BuildContext context, types.Message)? onMessageDoubleTap;
+
   /// Called when user makes a long press on any message
-  final void Function(types.Message)? onMessageLongPress;
+  final void Function(BuildContext context, types.Message)? onMessageLongPress;
 
   /// Called when user makes a long press on status icon in any message
-  final void Function(types.Message)? onMessageStatusLongPress;
+  final void Function(BuildContext context, types.Message)?
+      onMessageStatusLongPress;
 
   /// Called when user taps on status icon in any message
-  final void Function(types.Message)? onMessageStatusTap;
+  final void Function(BuildContext context, types.Message)? onMessageStatusTap;
 
   /// Called when user taps on any message
-  final void Function(types.Message)? onMessageTap;
+  final void Function(BuildContext context, types.Message)? onMessageTap;
+
+  /// Called when the message's visibility changes
+  final void Function(types.Message, bool visible)? onMessageVisibilityChanged;
 
   /// See [TextMessage.onPreviewDataFetched]
   final void Function(types.TextMessage, types.PreviewData)?
       onPreviewDataFetched;
+
+  /// See [TextMessage.previewTapOptions]
+  final PreviewTapOptions previewTapOptions;
 
   /// Rounds border of the message to visually group messages together.
   final bool roundBorder;
@@ -132,7 +152,7 @@ class Message extends StatelessWidget {
 
     return showAvatar
         ? Container(
-            margin: const EdgeInsets.only(right: 8),
+            margin: const EdgeInsetsDirectional.only(end: 8),
             child: GestureDetector(
               onTap: () => onAvatarTap?.call(message.author),
               child: CircleAvatar(
@@ -215,8 +235,10 @@ class Message extends StatelessWidget {
             : TextMessage(
                 emojiEnlargementBehavior: emojiEnlargementBehavior,
                 hideBackgroundOnEmojiMessages: hideBackgroundOnEmojiMessages,
+                isTextMessageTextSelectable: isTextMessageTextSelectable,
                 message: textMessage,
                 onPreviewDataFetched: onPreviewDataFetched,
+                previewTapOptions: previewTapOptions,
                 showName: showName,
                 usePreviewData: usePreviewData,
               );
@@ -284,32 +306,32 @@ class Message extends StatelessWidget {
                 emojiEnlargementBehavior, message as types.TextMessage);
     final _messageBorderRadius =
         InheritedChatTheme.of(context).theme.messageBorderRadius;
-    final _borderRadius = BorderRadius.only(
-      bottomLeft: Radius.circular(
+    final _borderRadius = BorderRadiusDirectional.only(
+      bottomEnd: Radius.circular(
+        _currentUserIsAuthor
+            ? roundBorder
+                ? _messageBorderRadius
+                : 0
+            : _messageBorderRadius,
+      ),
+      bottomStart: Radius.circular(
         _currentUserIsAuthor || roundBorder ? _messageBorderRadius : 0,
       ),
-      bottomRight: Radius.circular(_currentUserIsAuthor
-          ? roundBorder
-              ? _messageBorderRadius
-              : 0
-          : _messageBorderRadius),
-      topLeft: Radius.circular(_messageBorderRadius),
-      topRight: Radius.circular(_messageBorderRadius),
+      topEnd: Radius.circular(_messageBorderRadius),
+      topStart: Radius.circular(_messageBorderRadius),
     );
 
-    Alignment _getAlignMent() {
+    AlignmentDirectional _getAlignment() {
       return _currentUserIsAuthor
-          ? Alignment.centerRight
-          : Alignment.centerLeft;
+          ? AlignmentDirectional.centerEnd
+          : AlignmentDirectional.centerStart;
     }
 
     return Container(
-      alignment: messageAlignment != null
-          ? messageAlignment!(message) ?? _getAlignMent()
-          : _getAlignMent(),
-      margin: const EdgeInsets.only(
+      alignment: messageAlignment?.call(message) ?? _getAlignment(),
+      margin: const EdgeInsetsDirectional.only(
         bottom: 4,
-        left: 20,
+        start: 20,
       ),
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.end,
@@ -324,14 +346,28 @@ class Message extends StatelessWidget {
               crossAxisAlignment: CrossAxisAlignment.end,
               children: [
                 GestureDetector(
-                  onLongPress: () => onMessageLongPress?.call(message),
-                  onTap: () => onMessageTap?.call(message),
-                  child: _bubbleBuilder(
-                    context,
-                    _borderRadius,
-                    _currentUserIsAuthor,
-                    _enlargeEmojis,
-                  ),
+                  onDoubleTap: () => onMessageDoubleTap?.call(context, message),
+                  onLongPress: () => onMessageLongPress?.call(context, message),
+                  onTap: () => onMessageTap?.call(context, message),
+                  child: onMessageVisibilityChanged != null
+                      ? VisibilityDetector(
+                          key: Key(message.id),
+                          onVisibilityChanged: (visibilityInfo) =>
+                              onMessageVisibilityChanged!(message,
+                                  visibilityInfo.visibleFraction > 0.1),
+                          child: _bubbleBuilder(
+                            context,
+                            _borderRadius.resolve(Directionality.of(context)),
+                            _currentUserIsAuthor,
+                            _enlargeEmojis,
+                          ),
+                        )
+                      : _bubbleBuilder(
+                          context,
+                          _borderRadius.resolve(Directionality.of(context)),
+                          _currentUserIsAuthor,
+                          _enlargeEmojis,
+                        ),
                 ),
               ],
             ),
@@ -342,8 +378,8 @@ class Message extends StatelessWidget {
               child: showStatus
                   ? GestureDetector(
                       onLongPress: () =>
-                          onMessageStatusLongPress?.call(message),
-                      onTap: () => onMessageStatusTap?.call(message),
+                          onMessageStatusLongPress?.call(context, message),
+                      onTap: () => onMessageStatusTap?.call(context, message),
                       child: _statusBuilder(context),
                     )
                   : null,