+++ lib/src/widgets/send_button.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'inherited_chat_theme.dart';
 import 'inherited_l10n.dart';
 
@@ -8,27 +9,31 @@ class SendButton extends StatelessWidget {
   const SendButton({
     Key? key,
     required this.onPressed,
+    this.padding = EdgeInsets.zero,
   }) : super(key: key);
 
   /// Callback for send button tap event
   final void Function() onPressed;
 
+  /// Padding around the button
+  final EdgeInsets padding;
+
   @override
   Widget build(BuildContext context) {
     return Container(
-      height: 24,
-      margin: const EdgeInsets.only(left: 16),
-      width: 24,
+      margin: InheritedChatTheme.of(context).theme.sendButtonMargin ??
+          const EdgeInsetsDirectional.fromSTEB(0, 0, 8, 0),
       child: IconButton(
-        icon: InheritedChatTheme.of(context).theme.sendButtonIcon != null
-            ? InheritedChatTheme.of(context).theme.sendButtonIcon!
-            : Image.asset(
-                'assets/icon-send.png',
-                color: InheritedChatTheme.of(context).theme.inputTextColor,
-                package: 'flutter_chat_ui',
-              ),
+        constraints: const BoxConstraints(minHeight: 24, minWidth: 24),
+        icon: InheritedChatTheme.of(context).theme.sendButtonIcon ??
+            Image.asset(
+              'assets/icon-send.png',
+              color: InheritedChatTheme.of(context).theme.inputTextColor,
+              package: 'flutter_chat_ui',
+            ),
         onPressed: onPressed,
-        padding: EdgeInsets.zero,
+        padding: padding,
+        splashRadius: 24,
         tooltip: InheritedL10n.of(context).l10n.sendButtonAccessibilityLabel,
       ),
     );